# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: NodelistDB API
  description: |
    NodelistDB is a comprehensive API for accessing historical FTN nodelist
    data: every FidoNet nodelist since 1986, other FTN networks (fsxNet, ...),
    pointlists (FTS-5002), and what the testing daemon found when it called
    each node.

    ## Features
    - **Node search**: find nodes by zone, net, node number, system name,
      location, sysop, capability flags or date range, in any network
    - **Historical data**: the complete history, change log and timeline of
      any node
    - **Multi-network**: `GET /api/networks` lists the imported FTN networks;
      most endpoints take `?domain=`
    - **Points**: pointlist snapshots under a boss, point history, imported
      pointlist issues
    - **Statistics**: per-date network statistics and the list of nodelist dates
    - **Reachability**: the testing daemon's results per node (BinkP, IFCICO,
      Telnet, FTP, VModem, over IPv4 and IPv6), trends and searches
    - **Software, hosting, PING/TRACE analytics**: distributions computed from
      those tests, and the netmail PING measurements
    - **Flag documentation**: a reference for FidoNet node flags

    ## Multi-network conventions
    A node address is only unique within one FTN network. Endpoints about one
    address resolve the network in this order: an explicit `?domain=`; the
    only network the address exists in; `fidonet` when it exists in several.
...
    as their descriptions say.

...

...

...
servers:
...
paths:
...
  /api/nodes/{zone}/{net}/{node}/history:
    get:
      summary: Get Node History
      description: Every nodelist entry of one address, oldest first, with the span it covers.
      operationId: getNodeHistory
      tags:
        - Nodes
      parameters:
        - $ref: '#/components/parameters/Zone'
        - $ref: '#/components/parameters/Net'
        - $ref: '#/components/parameters/Node'
        - $ref: '#/components/parameters/DomainResolved'
      responses:
        '200':
          description: Node history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AddressEnvelope'
                  - type: object
                    properties:
                      history:
                        type: array
                        items:
                          $ref: '#/components/schemas/Node'
                      count:
                        type: integer
                      first_date:
                        type: string
                        format: date-time
                      last_date:
                        type: string
                        format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/reachability/trends:
    get:
      summary: Get Reachability Trends
      description: |
        Per day, how many nodes the daemon tested and how many were
        operational, with the average response time: the series behind the
        /reachability chart.
      operationId: getReachabilityTrends
      tags:
        - Reachability
      parameters:
        - $ref: '#/components/parameters/DomainDefaultAll'
        - name: days
          in: query
          description: Window in days; omitted or 0 means the whole history (values above 3650 are rejected)
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 3650
      responses:
        '200':
          description: Daily trend
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    type: string
                  days:
                    type: integer
                  trends:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReachabilityTrend'
                  count:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    AddressEnvelope:
      type: object
      description: >
        The preamble of every response about one 3-D or 4-D address: which
        network answered and which networks the address exists in.
      properties:
        address:
          type: string
          example: "2:5001/100"
        domain:
          type: string
          example: fidonet
        available_domains:
          type: array
          items:
            type: string
          description: Always an array, possibly empty
...
    Error:
      type: object
      description: Every JSON error body
      properties:
        error:
          type: string
          description: What went wrong
        status:
          type: integer
          description: The HTTP status, repeated
        time:
          type: string
          format: date-time
...
    Node:
      type: object
      properties:
        zone:
          type: integer
        net:
          type: integer
        node:
          type: integer
        domain:
          type: string
          description: Omitted when empty.
        nodelist_date:
          type: string
          format: date-time
        day_number:
          type: integer
        system_name:
          type: string
        location:
          type: string
        sysop_name:
          type: string
        phone:
          type: string
        node_type:
          type: string
        region:
          type: integer
          nullable: true
          description: Omitted when empty.
        max_speed:
          type: integer
        is_cm:
          type: boolean
        is_mo:
          type: boolean
        has_inet:
          type: boolean
        flags:
          type: array
          items:
            type: string
        modem_flags:
          type: array
          items:
            type: string
        internet_config:
          type: object
          additionalProperties: true
          description: Omitted when empty.
        conflict_sequence:
          type: integer
        has_conflict:
          type: boolean
        fts_id:
          type: string
        raw_line:
          type: string
          description: Omitted when empty.
...
    ReachabilityTrend:
      type: object
      properties:
        date:
          type: string
          format: date-time
        total_nodes:
          type: integer
        operational_nodes:
          type: integer
        failed_nodes:
          type: integer
        success_rate:
          type: number
        avg_response_ms:
          type: number
...
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        A caller-specific API key. The server stores its SHA-256 hash in
        modem_api.callers; see the README for generating a pair.
  responses:
    BadRequest:
      description: Invalid or missing parameter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No such record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: >
        Rate limited. The body is plain text and Retry-After says how many
        seconds to wait.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        text/plain:
          schema:
            type: string
    BudgetExceeded:
      description: The query outran its time budget; narrow it or retry later
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    Zone:
      name: zone
      in: path
      required: true
      description: FTN zone number (FidoNet uses 1-6; other networks have their own)
      schema:
        type: integer
        minimum: 1
    Net:
      name: net
      in: path
      required: true
      description: Network number (0-65535)
      schema:
        type: integer
        minimum: 0
        maximum: 65535
    Node:
      name: node
      in: path
      required: true
      description: Node number (0-65535)
      schema:
        type: integer
        minimum: 0
        maximum: 65535
...
    DomainResolved:
      name: domain
      in: query
      description: >
        FTN network. Omitted: the only network the address exists in, or
        fidonet when it exists in several.
      schema:
        type: string
        example: fsxnet
...
    DomainDefaultAll:
      name: domain
      in: query
      description: FTN network to report on (omitted = every network together)
      schema:
        type: string
...